Experiment/EventAIAssistant/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_community
langchain_openai
PyPDF2
faiss-cpu
langchain_google_genai
unstructured
jq
langchain_chroma
fastapi